requirements.txt
vertex-protocol-sdk
pandas
numpy
scikit-learn
